Rebasing, refactoring and fixing functional tests
