+++ test/e2e-test/features/basic.feature
@@ -18,7 +18,7 @@ Feature: Business rules
     When I click Your personal details
     Then I should be taken to the enter your name page
     When Enter "Random" "User" as my Given and Family Name and click Save and continue
-    Then click save and continue
+    Then I click Save and continue
     Then I should be taken to the DOB page
     When I enter "11" "11" "1999" as my DOB and click Save and continue
     Then I should be taken to nationality page
@@ -28,8 +28,8 @@ Feature: Business rules
     Then I should be taken to the what is your address page
     When I choose the first address from the dropdown list and click continue
     Then I should be taken to the confirm address page
-    When I confirm my address and click Save and continue
-    Then I should be taken to the task-list
+    And I click Save and continue
+    Then I should be taken to the task-list page
     Given I click the contact details link
     Then I should be taken to the contact-details page
     When I check the Text message option and type "07899999999" as my phone number and click Save and continue
@@ -37,7 +37,7 @@ Feature: Business rules
     When I click on the type-of-appeal link
     Then I should be taken to the appeal page
     When I click on Protection as my type of appeal and click Save and continue
-    Then I should be taken to the task-list
+    Then I should be taken to the task-list page
     When I click on the check and send your appeal link
     Then I should be taken to the check-and-send page
     # When I check the checkbox and click send
