+++ test/e2e-test/pages/contact-details/contact-details-page.ts
@@ -17,5 +17,19 @@ module.exports = {
     Then('I should see the contact details page', async () => {
       I.seeInCurrentUrl(paths.contactDetails);
     });
+
+    When(/^I check the Text message option and type "([^"]*)" as my phone number and click Save and continue$/, async (num) => {
+      await I.checkOption('Text message');
+      await I.fillField('#text-message-value', num);
+      await I.click('Save and continue');
+    });
+
+    Given(/^I click the contact details link$/, async () => {
+      await I.click('Your contact details');
+    });
+
+    Then(/^I should be taken to the contact\-details page$/, async () => {
+      await I.seeInCurrentUrl('contact-details');
+    });
   }
 };
