--- test/e2e-test/pages/home-office-details/ref-number-page.ts
@@ -1,39 +1,0 @@
-const { fillInDate } = require('../helper-functions');
-
-const { enterRefNumber } = require('../helper-functions');
-
-module.exports = {
-  fillInHomeOfficeDetails(I) {
-    When(/^I click on Home office details$/, async () => {
-      await I.click('a[href*="/home-office"]');
-    });
-
-    Then(/^I should be taken to the home office ref number page$/, async () => {
-      await I.seeInCurrentUrl('/home-office/details');
-    });
-
-    When(/^I enter "([^"]*)" as the Office ref number and click Save and continue/, async (refNumber) => {
-      enterRefNumber(refNumber);
-    });
-
-    Then(/^I should see letter sent page$/, async () => {
-      await I.seeInCurrentUrl('/letter-sent');
-    });
-
-    When(/^I enter an on time letter sent date and click Save and continue$/, async () => {
-      const date = new Date();
-      await fillInDate(date.getDate(),date.getMonth() + 1 ,date.getFullYear());
-      I.click('Save and continue');
-    });
-
-    When(/^I enter an out of time letter sent date and click Save and continue$/, async () => {
-      const date = new Date();
-      await fillInDate(date.getDate(),date.getMonth() + 1 ,date.getFullYear() - 1);
-      I.click('Save and continue');
-    });
-
-    Then(/^I expect to be redirect back to the task\-list$/, async () => {
-      await I.seeInCurrentUrl('/task-list');
-    });
-  }
-};
