+++ test/e2e-test/pages/personal-details/personal-details-date-of-birth.ts
@@ -1,4 +1,5 @@
 import { paths } from '../../../../app/paths';
+import { fillInDate } from '../helper-functions';
 const config = require('config');
 
 const testUrl = config.get('testUrl');
@@ -9,8 +10,17 @@ module.exports = {
       I.amOnPage(testUrl + paths.personalDetails.dob);
     });
 
+    Then(/^I should be taken to the DOB page$/, async () => {
+      await I.seeInCurrentUrl('details/date-of-birth');
+    });
+
     Then('I should see the date of birth page', async () => {
       I.seeInCurrentUrl(paths.personalDetails.dob);
+    });
+
+    When(/^I enter "([^"]*)" "([^"]*)" "([^"]*)" as my DOB and click Save and continue$/, (day,month,year) => {
+      fillInDate(day,month,year);
+      I.click('Save and continue');
     });
   }
 };
