+++ test/functional/features/home-office-letter-sent-page.feature
@@ -7,23 +7,23 @@ Feature: Home office letter sent page
     Given I have an appeal with home office reference
     And I have logged in
     And I am on the home office letter sent page
-    When I click save for later
+    When I click Save for later
     Then I should see the task-list page
     And I shouldnt be able to click "Your personal details"
 
     Given I am on the home office letter sent page
     When I enter a day "1" month "1" year "1"
-    And I click save for later
+    And I click Save for later
     Then I should see error summary
 
     Given I am on the home office letter sent page
     When I enter a day "1" month "1" year "1"
-    And I click save and continue
+    And I click Save and continue
     Then I should see error summary
 
     Given I am on the home office letter sent page
     When I enter a a home letter date in the last 2 weeks
-    And I click save for later
+    And I click Save for later
     Then I should see the task-list page
     And I should be able to click "Your personal details"
