+++ test/functional/features/personal-details-enter-postcode.feature
@@ -7,7 +7,7 @@ Scenario: Personal details enter postcode
 Given I have an appeal with home office details, name, date of birth and nationality
 And I have logged in
 And I am on the personal details enter postcode page
-When I click save for later
+When I click Save for later
 Then I should see the task-list page
 And I shouldnt be able to click "Your contact details"
