+++ test/functional/features/reason-for-appeal.feature
@@ -6,14 +6,14 @@ Feature: Reason for appeal
   Scenario: Navigate through reasons for appeal
     Given I am authenticated as a valid appellant
     When I visit reasons for appeal
-    Then I enter "" into the reason for appeal text box and click Save and Continue
+    And I click Save and continue
     Then I should see error summary
 
-    When I click save for later
+    When I click Save for later
     Then I should see error summary
 
     When I click the back button on reasons for appeal
-    Then I should be taken to the appllent timeline
+    Then I should be taken to the appellant timeline
 
     When I visit reasons for appeal
     Then I enter "Test" into the reason for appeal text box and click Save and Continue
