+++ test/unit/controllers/reason-for-appeal.test.ts
@@ -81,21 +81,22 @@ describe('Reasons for Appeal Controller', function() {
     it('should fail validation and render case-building/reasons-for-appeal/reason-for-appeal.njk with error', async () => {
       req.body.moreDetail = '';
       await postReasonForAppeal(updateAppealService as UpdateAppealService)(req as Request, res as Response, next);
+      const moreDetailError = {
+        text: 'Enter the reasons you think the Home Office decision is wrong',
+        href: '#moreDetail',
+        key: 'moreDetail'
+      };
+      const error = {
+        moreDetail: moreDetailError
+      };
+      const errorList = [ moreDetailError ];
+
       expect(res.render).to.have.been.calledWith(
-                'case-building/reasons-for-appeal/reason-for-appeal.njk',
+        'case-building/reasons-for-appeal/reason-for-appeal.njk',
         {
-          error: {
-            moreDetail: {
-              href: '#moreDetail',
-              key: 'moreDetail',
-              text: 'Enter the reasons you think the Home Office decision is wrong'
-            }
-          },
-          errorList: [{
-            href: '#moreDetail',
-            key: 'moreDetail',
-            text: 'Enter the reasons you think the Home Office decision is wrong'
-          }]
+          error,
+          errorList,
+          previousPage: '/appellant-timeline'
         }
       );
     });
